Refactor demo overview components to support multiple demo configurations and improve data handling
